chore(graphql): Move loaders to own folder
